+++ starter/critter/src/main/java/com/udacity/jdnd/course3/critter/dto/pet/PetDTO.java
@@ -1,4 +1,4 @@
-package com.udacity.jdnd.course3.critter.pet;
+package com.udacity.jdnd.course3.critter.dto.pet;
 
 import java.time.LocalDate;
