+++ starter/critter/src/main/java/com/udacity/jdnd/course3/critter/dto/users/CustomerDTO.java
@@ -1,4 +1,4 @@
-package com.udacity.jdnd.course3.critter.user;
+package com.udacity.jdnd.course3.critter.dto.users;
 
 import java.util.List;
