+++ starter/critter/src/main/java/com/udacity/jdnd/course3/critter/dto/users/EmployeeRequestDTO.java
@@ -1,4 +1,4 @@
-package com.udacity.jdnd.course3.critter.user;
+package com.udacity.jdnd.course3.critter.dto.users;
 
 import java.time.LocalDate;
 import java.util.Set;
