+++ starter/critter/src/main/java/com/udacity/jdnd/course3/critter/dto/users/EmployeeSkill.java
@@ -1,4 +1,4 @@
-package com.udacity.jdnd.course3.critter.user;
+package com.udacity.jdnd.course3.critter.dto.users;
 
 /**
  * A example list of employee skills that could be included on an employee or a schedule request.
