+++ starter/critter/src/main/resources/application.properties
@@ -4,7 +4,13 @@ server.port=8082
 # Data Source Setup
 # Note: Users must create a mysql db that they will reference here, create a user that has access to it on localhost
 #       and provide the credentials below to access it
+spring.datasource.url=jdbc:mysql://localhost:3306/plant
+spring.datasource.username=sa
+spring.datasource.password=sa
 
 # uncomment the below lines to view sql output in console
-#spring.jpa.show-sql=true
-#spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+
+spring.datasource.initialization-mode=always
+spring.jpa.hibernate.ddl-auto=create
