+++ starter/critter/src/test/java/com/udacity/jdnd/course3/critter/CritterFunctionalTest.java
@@ -2,12 +2,16 @@ package com.udacity.jdnd.course3.critter;
 
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
-import com.udacity.jdnd.course3.critter.pet.PetController;
-import com.udacity.jdnd.course3.critter.pet.PetDTO;
-import com.udacity.jdnd.course3.critter.pet.PetType;
-import com.udacity.jdnd.course3.critter.schedule.ScheduleController;
-import com.udacity.jdnd.course3.critter.schedule.ScheduleDTO;
-import com.udacity.jdnd.course3.critter.user.*;
+import com.udacity.jdnd.course3.critter.controller.PetController;
+import com.udacity.jdnd.course3.critter.controller.UserController;
+import com.udacity.jdnd.course3.critter.dto.pet.PetDTO;
+import com.udacity.jdnd.course3.critter.dto.pet.PetType;
+import com.udacity.jdnd.course3.critter.controller.ScheduleController;
+import com.udacity.jdnd.course3.critter.dto.schedule.ScheduleDTO;
+import com.udacity.jdnd.course3.critter.dto.users.CustomerDTO;
+import com.udacity.jdnd.course3.critter.dto.users.EmployeeDTO;
+import com.udacity.jdnd.course3.critter.dto.users.EmployeeRequestDTO;
+import com.udacity.jdnd.course3.critter.dto.users.EmployeeSkill;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
